+++ tools/debug.py
@@ -1,0 +1,3 @@
+from torrt.main import process_commands
+
+process_commands(['register_torrent', '123', '--params="a=b d=c"'])